kafka-tools/kafka-topic-convert/requirements.txt
pyaml>=20.0
kafka-python>=2.0
